Move core functions of database to __init__
For example, it's more reasonable to have `get_database` available
with `from database import get_database` instead of `from database.util
import ...` since it's the core function.
